Move estimator to own class
